+++ src/features/auth/form.hook.ts
@@ -1,0 +1,65 @@
+import { yupResolver } from '@hookform/resolvers/yup';
+import { useEffect, useState } from 'react';
+import { useForm } from 'react-hook-form';
+
+import { JOB_CATEGORIES, type FormValues } from './form.types';
+import { formConfig } from './form.utils';
+
+export const useProfileForm = () => {
+  const [formStructure, setFormStructure] = useState(formConfig.structure);
+  const method = useForm<FormValues>({
+    resolver: yupResolver(formConfig.validation),
+    mode: 'onChange',
+    defaultValues: {
+      name: '',
+      briefIntro: '',
+      majorJobGroup: null,
+      minorJobGroup: null,
+      jobTitle: '',
+      division: 'student',
+      url: [{ value: '' }],
+      imageUrl: '',
+    },
+  });
+  const majorJobGroup = method.watch('majorJobGroup');
+
+  /**
+   * 직군 선택 시 직무 불러오기
+   * majorJobGroup = Option ({value, label}), 직군
+   * formConfig의 minorJobGroup input option에 직무 그룹 추가
+   */
+  useEffect(() => {
+    if (!majorJobGroup) return;
+
+    setFormStructure(prev => {
+      const updatedStructure = [...prev];
+
+      for (let i = 0; i < updatedStructure.length; i++) {
+        let findIdx;
+        const { inputs } = updatedStructure[i];
+        if ((findIdx = inputs.findIndex(input => input.name === 'minorJobGroup')) !== -1) {
+          const minorJobGroup = [...JOB_CATEGORIES].find(
+            category => category.value === majorJobGroup.value,
+          )?.children;
+          inputs[findIdx].options = minorJobGroup;
+          inputs[findIdx].placeholder = '직무를 선택해 주세요.';
+          break;
+        }
+      }
+
+      return updatedStructure;
+    });
+
+    method.setValue('minorJobGroup', null, { shouldValidate: true });
+  }, [majorJobGroup, method]);
+
+  const onSubmit = (data: FormValues) => {
+    console.log('폼 데이터:', data);
+  };
+
+  return {
+    formStructure,
+    method,
+    onSubmit,
+  };
+};
